backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12
email-validator==2.2.0
python-dotenv==1.0.1